--- src/utils/hasPermission.tsx
@@ -1,6 +1,0 @@
-export enum Permission {
-  ViewHome = 'viewHome',
-  ViewUserProfile = 'viewUserProfile',
-  EditUser = 'editUser',
-  EditSubject = 'editSubject',
-}
